requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentence-transformers
faiss-cpu
numpy
pydantic
python-multipart

# Mistral 7B 4-bit on GPU
accelerate
bitsandbytes

# Streaming + fallback + eval
requests
rapidfuzz
